move `build_c_file` and `build_rust_file` into `SupportedArchitectureTest`
